hotfix : Confirm Delete 1. Confirm Post Delete 2. Confirm Comment Delete

+++ src/components/comment/Comment.tsx
@@ -13,7 +13,6 @@ import CommentList from './CommentList';
 const Comment = () => {
   const { pathname } = useLocation();
   const { id: boardId } = useParams();
-  const navigate = useNavigate();
   const isRootPath = pathname === ROUTE_PATH.ROOT;
   const skipQuery = { skip: isRootPath || !boardId };
   const { data: boardList, isFetching: boardFetching } = useGetBoardByIdQuery(
@@ -31,7 +30,9 @@ const Comment = () => {
     !!commentList?.length && !!boardList?.content?.defense.length;
 
   const handleDeleteComment = async (commentId: number) => {
-    await deleteComment(commentId).then(refetch);
+    if (window.confirm('Are you sure you want to delete this item?')) {
+      await deleteComment(commentId).then(refetch);
+    }
   };
 
   return (
